% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_peak.R
\name{estimate_Genome_peakp}
\alias{estimate_Genome_peakp}
\title{Function to fit 2p peak model, with p forms}
\usage{
estimate_Genome_peakp(kmer_hist_orig, x, y, k, p, topology, estKmercov,
  round, foldername, arguments)
}
\arguments{
\item{kmer_hist_orig}{A data frame of the original histogram data (starting at 1 and with last position removed).}

\item{x}{An integer vector of the x-coordinates of the histogram (after filtering out low coverage errors and high coverage kmers).}

\item{y}{A numeric vector of the y-coordinates of the histogram (after filtering out low coverage errors and high coverage kmers).}

\item{k}{An integer corresponding to the kmer length.}

\item{p}{An integer corresponding to the ploidy.}

\item{topology}{An integer corresponding to the topology to use.}

\item{estKmercov}{An integer corresponding to the estimated kmer coverage of the polyploid genome.
Set to -1 if not specified by user.}

\item{round}{An integer corresponding to the iteration number (0, 1, 2, 3) for the fitting process.}

\item{foldername}{A character vector corresponding to the name of the output directory.}

\item{arguments}{A data frame of the user-specified inputs.}
}
\value{
A list (nls, nlsscore) where nls is the nlsLM model object (with some additional components)
and nlsscore is the score (model RSSE) corresponding to the best fit (of the p forms).
}
\description{
Function to fit 2p peak model, with p forms
}
