% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{eval_model}
\alias{eval_model}
\title{Evaluate distinct model forms, in order to resolve ambiguity of which peak is the homozygous peak}
\usage{
eval_model(kmer_hist_orig, nls0, nls1, p, round, foldername, arguments)
}
\arguments{
\item{kmer_hist_orig}{A data frame of the original histogram data (starting at 1 and with last position removed).}

\item{nls0, nls1}{The nlsLM model objects to evaluate and compare.}

\item{p}{An integer corresponding to the ploidy.}

\item{round}{An integer corresponding to the iteration number (0, 1, 2, 3) for the fitting process.}

\item{foldername}{A character vector corresponding to the name of the output directory.}

\item{arguments}{A data frame of the user-specified inputs.}
}
\value{
A list (nls, nlsscore) where nls is the nlsLM model object (with some additional components)
and nlsscore is the score (model RSSE) corresponding to the best fit (of the p forms).
}
\description{
Evaluate distinct model forms, in order to resolve ambiguity of which peak is the homozygous peak
}
