% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_peak.R
\name{nls_peak}
\alias{nls_peak}
\title{Uses nlsLM to fit 2p peak model}
\usage{
nls_peak(x, y, k, p, estKmercov, estLength, max_iterations)
}
\arguments{
\item{x}{An integer vector of the x-coordinates of the histogram (after filtering out low coverage errors and high coverage kmers).}

\item{y}{A numeric vector of the y-coordinates of the histogram (after filtering out low coverage errors and high coverage kmers).}

\item{k}{An integer corresponding to the kmer length.}

\item{p}{An integer corresponding to the ploidy.}

\item{estKmercov}{A numeric corresponding to the estimated average kmer coverage of the polyploid genome.}

\item{estLength}{A numeric corresponding to the estimated polyploid genome length.}

\item{max_iterations}{An integer corresponding to the maximum number iterations to use for nlsLM.}
}
\value{
An nlsLM model object with some additional components.
}
\description{
Uses nlsLM to fit 2p peak model
}
