% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict1_unique}
\alias{predict1_unique}
\title{something}
\usage{
predict1_unique(k, d, kmercov, bias, x)
}
\value{
something
}
\description{
something
}
\examples{
blah blah blah
}
\seealso{
something
}
