% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict2_unique}
\alias{predict2_unique}
\title{something}
\usage{
predict2_unique(r1, k, d, kmercov, bias, x)
}
\value{
something
}
\description{
something
}
\examples{
blah blah blah
}
\seealso{
something
}
