% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict3_unique}
\alias{predict3_unique}
\title{something}
\usage{
predict3_unique(r1, r2, k, d, kmercov, bias, x)
}
\value{
something
}
\description{
something
}
\examples{
blah blah blah
}
\seealso{
something
}
