% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict4_unique}
\alias{predict4_unique}
\title{something}
\usage{
predict4_unique(r1, r2, r3, k, d, kmercov, bias, x)

predict4_unique(r1, r2, r3, k, d, kmercov, bias, x)
}
\value{
something

something
}
\description{
something

something
}
\examples{
blah blah blah
blah blah blah
}
\seealso{
something

something
}
