% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict5}
\alias{predict5}
\title{something}
\usage{
predict5(r1, r2, r3, r4, r5, r6, k, d, kmercov, bias, x)
}
\value{
something
}
\description{
something
}
\examples{
blah blah blah
}
\seealso{
something
}
