% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict6_5}
\alias{predict6_5}
\title{Produce model estimated (p=6, topology=5) y-coordinates of the kmer spectra given the kmer size, repetitiveness, average polyploid kmer coverage, bias, and x-coordinates of the kmer spectra.}
\usage{
predict6_5(raaaaab, raaabbc, raabccd, rabcdde, rabcdef, k, d, kmercov,
  bias, x)
}
\arguments{
\item{raaaaab, raaabbc, raabccd, rabcdde, rabcdef}{Numerics corresponding to the nucleotide heterozygosities aaaaab, aaabbc, aabccd, abcdde, and abcdef respectively.}

\item{k}{An integer corresponding to the kmer length.}

\item{d}{A numeric corresponding to the repetitiveness.}

\item{kmercov}{A numeric corresponding to the estimated average kmer coverage of the polyploid genome.}

\item{bias}{A numeric corresponding to the overdispersion of the negative binomial distribution.}

\item{x}{An integer vector of the x-coordinates of the histogram (after filtering out low coverage errors and high coverage kmers).}
}
\value{
A numeric vector of the model estimated y-coordinates of the kmer spectra.
}
\description{
Produce model estimated (p=6, topology=5) y-coordinates of the kmer spectra given the kmer size, repetitiveness, average polyploid kmer coverage, bias, and x-coordinates of the kmer spectra.
}
