% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict6_unique}
\alias{predict6_unique}
\title{something}
\usage{
predict6_unique(r1, r2, r3, r4, r5, r6, r7, r8, r9, r10, k, d, kmercov,
  bias, x)
}
\value{
something
}
\description{
something
}
\examples{
blah blah blah
}
\seealso{
something
}
