% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{report_results}
\alias{report_results}
\title{Report results and make plots}
\usage{
report_results(kmer_hist, kmer_hist_orig, k, p, container, foldername,
  arguments, IN_VERBOSE)
}
\arguments{
\item{kmer_hist}{A data frame of the original histogram data (starting at 1 and going up to the max kmer coverage threshold).}

\item{kmer_hist_orig}{A data frame of the original histogram data (starting at 1 and with last position removed).}

\item{k}{An integer corresponding to the kmer length.}

\item{p}{An integer corresponding to the ploidy.}

\item{container}{A list (nls, nlsscore) where nls is the nlsLM model object (with some additional components)
and nlsscore is the score (model RSSE) corresponding to the best fit (of the p forms).}

\item{foldername}{A character vector corresponding to the name of the output directory.}

\item{arguments}{A data frame of the user-specified inputs.}

\item{IN_VERBOSE}{A boolean flag to designate whether report_results is being called in a VERBOSE block.}
}
\description{
Report results and make plots
}
