% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_model.R
\name{score_model}
\alias{score_model}
\title{Score nlsLM model by number and percent of residual errors after excluding sequencing errors}
\usage{
score_model(kmer_hist_orig, nls, round, foldername)
}
\arguments{
\item{kmer_hist_orig}{A data frame of the original histogram data (starting at 1 and with last position removed).}

\item{nls}{An nlsLM model object.}

\item{round}{An integer corresponding to the iteration number (0, 1, 2, 3) for the fitting process.}

\item{foldername}{A character vector corresponding to the name of the output directory.}
}
\value{
A data frame where the variables "all", "full", and "unique" correspond to the model residual sum of square error (excluding sequencing error)
including x-values up to the end of kmer_hist_orig, up to (2p+1)*lambda, and up to (p+1)*lambda respectively.
Additionally, the variables "allscore", "fullscore", and "uniquescore" are the
corresponding percentage of kmers that are correctly modeled.
}
\description{
Score nlsLM model by number and percent of residual errors after excluding sequencing errors
}
